Se refactoriza el flujo de customer

+++ controllers/addressesController.js
@@ -57,6 +57,10 @@ const getAddressByCustomerId = async (req, res) => {
 const deleteAddress = async (req, res) => {
     const { id } = req.params;
     try {
+        const addressIsPrimary = await db.query('SELECT is_primary FROM YuNowDataBase.addresses WHERE id = $1', [id]);
+        if (addressIsPrimary.rows.length ===1){
+            return res.status(404).json({ error: 'No se puede eliminar la dirección ya que es primaria.' });
+        }
         const result = await db.query('DELETE FROM YuNowDataBase.addresses WHERE id = $1 RETURNING *', [id]);
 
         if (result.rows.length === 0) {
